数据统计

+++ inbyte-component/inbyte-component-admin/inbyte-component-admin-user/src/main/java/com/inbyte/component/admin/user/model/data/UserWeixinStat.java
@@ -3,13 +3,15 @@ package com.inbyte.component.admin.user.model.data;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class UserWeixinStat {
     private int totalUsers;
-    private double dailyGrowth;
-    private double weeklyGrowth;
-    private double monthlyGrowth;
+    private BigDecimal dailyGrowth;
+    private BigDecimal weeklyGrowth;
+    private BigDecimal monthlyGrowth;
     private int newUsers;
     private int message;
     private int money;
